requirements.txt
openai>=1.30
pdf2image>=1.17
Pillow>=10.0
openpyxl>=3.1
tqdm>=4.66
tenacity>=8.2
python-dotenv>=1.0
opencv-python>=4.8
numpy>=1.24
